Update main.py
I was concerned about timers creeping backwards over time.  For example, if you set every 12 hours then you're fine, because every 12 hours it will run.  If you set every 11:59:59 though then it will be one minutes back (as it runs every 60 seconds) each time, and after a number of cycles you end up way off track.

Added a daily reset of the array, since the maximum a timer can be is 24 hours I figure resetting it at midnight daily is safe, that means all timers are set based off of midnight (or 12:01 technically) and the potential for them to creep is minimized.